backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.6.3
packaging==24.0
python-multipart==0.0.26
groq==1.0.0
sentence-transformers==5.1.2
pypdf==5.1.0
duckduckgo-search==6.2.13
certifi
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
